Complete exercise 5.13 (Site scraper)
Modify a previous example that does a breadth-first search of links on a
specified page so that it saves the contents of the page, copying the
directory structure.

Doesn't work perfectly: If a link is a directory but doesn't end with a
slash, it will save it as a file with that name. If it continues into
that path, all files within it will fail, since the directory name has
already been created as a file.